+++ main/core/Resources/modules/html5-touch-backend/HTML5TouchBackend.js
