+++ main/core/Resources/modules/html5-touch-backend/index.js
