+++ plugin/exo/Resources/modules/quiz/quiz.js
@@ -15,7 +15,7 @@ import {MODAL_ADD_ITEM, AddItemModal} from './editor/components/add-item-modal.j
 import {MODAL_IMPORT_ITEMS, ImportItemsModal} from './editor/components/import-items-modal.jsx'
 import {MODAL_ADD_CONTENT, AddContentModal} from './editor/components/add-content-modal.jsx'
 import {MODAL_CONTENT, ContentModal} from './../contents/components/content-modal.jsx'
-import {default as HTML5Backend} from './../lib/HTML5TouchBackend/index'
+import {default as HTML5Backend} from '#/main/core/html5-touch-backend'
 
 export class Quiz {
   constructor(rawQuizData, noServer = false) {
